fix test and some refactoring
